backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.1
pydantic==2.10.3
pandas==2.2.3
reportlab==4.2.5
python-multipart==0.0.20
stripe==11.2.0
python-dateutil==2.9.0
